backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
django==5.2.9
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
environs==14.5.0
gunicorn==23.0.0
idna==3.11
marshmallow==4.1.1
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
pyjwt==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
urllib3==2.6.2
